fix(icon): WizIKeepBold
