+++ packages/wiz-ui-next/src/components/icons/index.ts
@@ -83,7 +83,7 @@ import { default as WizIInboxCustomize } from "./inbox-customize.vue";
 import { default as WizIInfoBold } from "./info-bold.vue";
 import { default as WizIInfoFilled } from "./info-filled.vue";
 import { default as WizIInfo } from "./info.vue";
-import { default as WizIKeep } from "./keep.vue";
+import { default as WizIKeepBold } from "./keep-bold.vue";
 import { default as WizILinkOff } from "./link-off.vue";
 import { default as WizILocationCity } from "./location-city.vue";
 import { default as WizILocationOn } from "./location-on.vue";
@@ -208,6 +208,7 @@ export type TIcon =
   | typeof WizIInfoBold
   | typeof WizIInfoFilled
   | typeof WizIInfo
+  | typeof WizIKeepBold
   | typeof WizIKeep
   | typeof WizILinkOff
   | typeof WizILocationCity
@@ -248,18 +249,18 @@ export type TIcon =
   | typeof WizIWarning;
 
 export {
-  WizIAddCircle,
   WizIAdd,
-  WizIAdjustBold,
+  WizIAddCircle,
   WizIAdjust,
+  WizIAdjustBold,
   WizIApprovalFilled,
   WizIArrowDownward,
   WizIArrowDropDown,
   WizIArrowDropUp,
   WizIArrowRight,
   WizIArrowUpward,
-  WizIAssignmentAdd,
   WizIAssignment,
+  WizIAssignmentAdd,
   WizIAttachFile,
   WizIAutoRenew,
   WizIBarChart4Bars,
@@ -270,41 +271,41 @@ export {
   WizICamera,
   WizICancel,
   WizIChangeHistory,
-  WizIChatBubbleFilled,
   WizIChat,
+  WizIChatBubbleFilled,
+  WizICheck,
   WizICheckBold,
   WizICheckCircleFilled,
-  WizICheck,
-  WizIChevronLeftBold,
   WizIChevronLeft,
-  WizIChevronRightBold,
+  WizIChevronLeftBold,
   WizIChevronRight,
-  WizICircleCheck,
+  WizIChevronRightBold,
   WizICircle,
-  WizICloseBold,
+  WizICircleCheck,
   WizIClose,
+  WizICloseBold,
   WizICloudUpload,
   WizIContactPageFilled,
   WizIContentCopy,
   WizIContract,
-  WizICounter1Filled,
   WizICounter1,
-  WizICounter2Filled,
+  WizICounter1Filled,
   WizICounter2,
-  WizICounter3Filled,
+  WizICounter2Filled,
   WizICounter3,
-  WizICounter4Filled,
+  WizICounter3Filled,
   WizICounter4,
-  WizICounter5Filled,
+  WizICounter4Filled,
   WizICounter5,
-  WizICounter6Filled,
+  WizICounter5Filled,
   WizICounter6,
-  WizICounter7Filled,
+  WizICounter6Filled,
   WizICounter7,
-  WizICounter8Filled,
+  WizICounter7Filled,
   WizICounter8,
-  WizICounter9Filled,
+  WizICounter8Filled,
   WizICounter9,
+  WizICounter9Filled,
   WizIDashboard,
   WizIDelete,
   WizIDemography,
@@ -317,8 +318,8 @@ export {
   WizIEmergencyHome,
   WizIErrorFilled,
   WizIExpandLess,
-  WizIExpandMoreBold,
   WizIExpandMore,
+  WizIExpandMoreBold,
   WizIEye,
   WizIFamilyHome,
   WizIFeedback,
@@ -330,18 +331,19 @@ export {
   WizIHourglassTop,
   WizIImageBold,
   WizIInboxCustomize,
+  WizIInfo,
   WizIInfoBold,
   WizIInfoFilled,
-  WizIInfo,
   WizIKeep,
+  WizIKeepBold,
   WizILinkOff,
   WizILocationCity,
   WizILocationOn,
   WizILogin,
   WizIMail,
   WizIManageAccounts,
-  WizIMenuOpen,
   WizIMenu,
+  WizIMenuOpen,
   WizIMonitoring,
   WizIMoreVert,
   WizINoteAdd,
@@ -353,8 +355,8 @@ export {
   WizIPinDrop,
   WizIPlayCircleBold,
   WizIPublic,
-  WizIRemoveBold,
   WizIRemove,
+  WizIRemoveBold,
   WizIReviews,
   WizISchedule,
   WizISearch,
