--- packages/wiz-ui-next/src/components/icons/keep.vue
@@ -1,20 +1,0 @@
-<template>
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    height="1em"
-    viewBox="0 -960 960 960"
-    width="1em"
-  >
-    <path
-      d="M600-760v296.92l57.23 57.23q2.92 2.93 4.39 6.58 1.46 3.65 1.46 7.55v11.21q0 8.28-5.97 14.39-5.98 6.12-14.8 6.12H500v197.08q0 8.5-5.76 14.25t-14.27 5.75q-8.51 0-14.24-5.75T460-162.92V-360H317.69q-8.82 0-14.8-6.12-5.97-6.11-5.97-14.39v-11.21q0-3.9 1.46-7.55 1.47-3.65 4.39-6.58L360-463.08V-760h-20q-8.5 0-14.25-5.76T320-780.03q0-8.51 5.75-14.24T340-800h280q8.5 0 14.25 5.76t5.75 14.27q0 8.51-5.75 14.24T620-760h-20ZM354-400h252l-46-46v-314H400v314l-46 46Zm126 0Z"
-    />
-  </svg>
-</template>
-
-<script setup lang="ts">
-import { ComponentName } from "@wizleap-inc/wiz-ui-constants";
-
-defineOptions({
-  name: ComponentName.IKeep,
-});
-</script>
