+++ packages/wiz-ui-react/src/components/icons/index.ts
@@ -83,7 +83,7 @@ import { WizIInboxCustomize } from "./inbox-customize";
 import { WizIInfo } from "./info";
 import { WizIInfoBold } from "./info-bold";
 import { WizIInfoFilled } from "./info-filled";
-import { WizIKeep } from "./keep";
+import { WizIKeepBold } from "./keep-bold";
 import { WizILinkOff } from "./link-off";
 import { WizILocationCity } from "./location-city";
 import { WizILocationOn } from "./location-on";
@@ -208,6 +208,7 @@ export type TIcon =
   | typeof WizIInfoBold
   | typeof WizIInfoFilled
   | typeof WizIInfo
+  | typeof WizIKeepBold
   | typeof WizIKeep
   | typeof WizILinkOff
   | typeof WizILocationCity
@@ -248,18 +249,18 @@ export type TIcon =
   | typeof WizIWarning;
 
 export {
-  WizIAddCircle,
   WizIAdd,
-  WizIAdjustBold,
+  WizIAddCircle,
   WizIAdjust,
+  WizIAdjustBold,
   WizIApprovalFilled,
   WizIArrowDownward,
   WizIArrowDropDown,
   WizIArrowDropUp,
   WizIArrowRight,
   WizIArrowUpward,
-  WizIAssignmentAdd,
   WizIAssignment,
+  WizIAssignmentAdd,
   WizIAttachFile,
   WizIAutoRenew,
   WizIBarChart4Bars,
@@ -270,41 +271,41 @@ export {
   WizICamera,
   WizICancel,
   WizIChangeHistory,
-  WizIChatBubbleFilled,
   WizIChat,
+  WizIChatBubbleFilled,
+  WizICheck,
   WizICheckBold,
   WizICheckCircleFilled,
-  WizICheck,
-  WizIChevronLeftBold,
   WizIChevronLeft,
-  WizIChevronRightBold,
+  WizIChevronLeftBold,
   WizIChevronRight,
-  WizICircleCheck,
+  WizIChevronRightBold,
   WizICircle,
-  WizICloseBold,
+  WizICircleCheck,
   WizIClose,
+  WizICloseBold,
   WizICloudUpload,
   WizIContactPageFilled,
   WizIContentCopy,
   WizIContract,
-  WizICounter1Filled,
   WizICounter1,
-  WizICounter2Filled,
+  WizICounter1Filled,
   WizICounter2,
-  WizICounter3Filled,
+  WizICounter2Filled,
   WizICounter3,
-  WizICounter4Filled,
+  WizICounter3Filled,
   WizICounter4,
-  WizICounter5Filled,
+  WizICounter4Filled,
   WizICounter5,
-  WizICounter6Filled,
+  WizICounter5Filled,
   WizICounter6,
-  WizICounter7Filled,
+  WizICounter6Filled,
   WizICounter7,
-  WizICounter8Filled,
+  WizICounter7Filled,
   WizICounter8,
-  WizICounter9Filled,
+  WizICounter8Filled,
   WizICounter9,
+  WizICounter9Filled,
   WizIDashboard,
   WizIDelete,
   WizIDemography,
@@ -317,8 +318,8 @@ export {
   WizIEmergencyHome,
   WizIErrorFilled,
   WizIExpandLess,
-  WizIExpandMoreBold,
   WizIExpandMore,
+  WizIExpandMoreBold,
   WizIEye,
   WizIFamilyHome,
   WizIFeedback,
@@ -330,18 +331,19 @@ export {
   WizIHourglassTop,
   WizIImageBold,
   WizIInboxCustomize,
+  WizIInfo,
   WizIInfoBold,
   WizIInfoFilled,
-  WizIInfo,
   WizIKeep,
+  WizIKeepBold,
   WizILinkOff,
   WizILocationCity,
   WizILocationOn,
   WizILogin,
   WizIMail,
   WizIManageAccounts,
-  WizIMenuOpen,
   WizIMenu,
+  WizIMenuOpen,
   WizIMonitoring,
   WizIMoreVert,
   WizINoteAdd,
@@ -353,8 +355,8 @@ export {
   WizIPinDrop,
   WizIPlayCircleBold,
   WizIPublic,
-  WizIRemoveBold,
   WizIRemove,
+  WizIRemoveBold,
   WizIReviews,
   WizISchedule,
   WizISearch,
